+++ package/rhythmicon-rhythm/CHANGES.md
@@ -1,3 +1,10 @@
+# 0.4.0
+
+- Make `0` a rest value as well
+- Add factory methods `fromTracy` and `fromHex`
+- Rename factory method `euclidean` to `fromEuclidean`
+- Add method `includes`
+
 # 0.3.0
 
 - Disallow `|...|` pattern strings